clean code new
